pyproject.toml
[project]
name = "sporty-api-tests"
version = "0.1.0"
description = ""
authors = [
    {name = "Martin Stajnko"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest-playwright (>=0.7.1,<0.8.0)",
    "playwright (>=1.55.0,<2.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
